move DictKeyType defn basecontainer.py -> base.py
